refactor: Reorganize file structure and improve error handling in database management actions
- The project's file structure has been reorganized, moving all database management action files to the actions/manage directory, improving code organization and making it easier to navigate.
- Improved error handling in database management actions, providing more detailed error information and improving application stability.

+++ components/Review/Review.js
@@ -1,13 +1,8 @@
 "use client";
 
 import { useState } from "react";
-import { useToast } from "@/hooks/useToast";
-
-import Dropdown from "../Dropdown";
-import ReviewForm from "./ReviewForm";
-
-import { deleteReview } from "@/actions/reviewManager";
-import { ReviewRating } from "./ReviewRating";
+import toast from "react-hot-toast";
+import useConfirm from "@/hooks/useConfirm";
 
 import {
   AlertDialog,
@@ -20,9 +15,14 @@ import {
   AlertDialogTitle,
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
+import Dropdown from "../Dropdown";
+import ReviewForm from "./ReviewForm";
+import { ReviewRating } from "./ReviewRating";
+
+import { deleteReview } from "@/actions/menage/reviewManager";
 
 const Review = ({ comment }) => {
-  const toast = useToast();
+  const { confirm } = useConfirm();
   const deleteReviewWithId = deleteReview.bind(null, comment.id);
 
   const [isEditing, setIsEditing] = useState(false);
@@ -39,17 +39,22 @@ const Review = ({ comment }) => {
     setIsEditing(false);
   };
 
-  const handleDeleteComment = async () => {
-    try {
-      const response = await deleteReviewWithId();
+  const handleDeleteReview = async () => {
+    const isConfirmed = await confirm(
+      "Tego działania nie można cofnąć. Spowoduje to trwałe usunięcie Twojej opinii z naszych serwerów"
+    );
 
-      if (response?.errors) {
-        toast.error(response.message);
-      } else {
-        toast.success(response.message);
+    if (isConfirmed) {
+      try {
+        const result = await deleteReviewWithId();
+        if (result.success === true) {
+          toast.success(result.message);
+        } else if (result.success === false) {
+          toast.error(result.message);
+        }
+      } catch (error) {
+        toast.error(error.message);
       }
-    } catch (error) {
-      toast.error("Wystąpił błąd podczas usuwania komentarza.");
     }
   };
 
@@ -143,48 +148,28 @@ const Review = ({ comment }) => {
                       <span className="mx-1">Edycja</span>
                     </button>
 
-                    <AlertDialog>
-                      <AlertDialogTrigger asChild>
-                        <button
-                          type="button"
-                          className="flex w-full items-center p-3 text-sm text-gray-600 capitalize transition-colors duration-300 transform hover:bg-gray-100"
-                        >
-                          <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            fill="none"
-                            viewBox="0 0 24 24"
-                            strokeWidth="1.5"
-                            stroke="currentColor"
-                            className="w-5 h-5 mx-1"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
-                            />
-                          </svg>
+                    <button
+                      onClick={handleDeleteReview}
+                      type="button"
+                      className="flex w-full items-center p-3 text-sm text-gray-600 capitalize transition-colors duration-300 transform hover:bg-gray-100"
+                    >
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        fill="none"
+                        viewBox="0 0 24 24"
+                        strokeWidth="1.5"
+                        stroke="currentColor"
+                        className="w-5 h-5 mx-1"
+                      >
+                        <path
+                          strokeLinecap="round"
+                          strokeLinejoin="round"
+                          d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
+                        />
+                      </svg>
 
-                          <span className="mx-1">Usuń</span>
-                        </button>
-                      </AlertDialogTrigger>
-                      <AlertDialogContent>
-                        <AlertDialogHeader>
-                          <AlertDialogTitle>
-                            Czy jesteś całkowicie pewien?
-                          </AlertDialogTitle>
-                          <AlertDialogDescription>
-                            Tego działania nie można cofnąć. Spowoduje to trwałe
-                            usunięcie Twojej opinii z naszych serwerów.
-                          </AlertDialogDescription>
-                        </AlertDialogHeader>
-                        <AlertDialogFooter>
-                          <AlertDialogCancel>Anuluj</AlertDialogCancel>
-                          <AlertDialogAction onClick={handleDeleteComment}>
-                            Kontynuuj
-                          </AlertDialogAction>
-                        </AlertDialogFooter>
-                      </AlertDialogContent>
-                    </AlertDialog>
+                      <span className="mx-1">Usuń</span>
+                    </button>
                   </>
                 }
               />
